fix: handle set_active_configuration failure() in rediscover()
set_active_configuration() may fail with rusb::Error::Busy, which means
interfaces are currently claimed. This is a known issue with libusb: "When
libusb presents a device handle to an application, there is a chance that the
corresponding device may be in unconfigured state."
(https://libusb.sourceforge.io/api-1.0/libusb_caveats.html)

This issue happens mostly on Linux. To work around this, we retry the operation
a few times. It's not perfect but efficient.